Refactor of the changes made before.
