+++ activity_browser/bwutils/graph_traversal/__init__.py
@@ -1,0 +1,5 @@
+"""Brightway graph-traversal helpers for the Tree and Sankey LCA Results tabs.
+
+Import from ``engine``, ``tree``, ``sankey``, or ``partition_plots``.
+This package does not re-export symbols.
+"""
